quiz/requirements.txt
boto3
langchain_core==0.3.40
langchain_google_genai==2.0.10
google-api-python-client==2.162.0
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.4
python-dotenv==1.0.1
